Updated tests and made them more generic and thorough
